Moved placeholder data to separate folder
